requirements.txt
# Core dependencies for Congressional Trading Monitor

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0

# PDF processing
pdfplumber>=0.9.0

# Retry logic and reliability
tenacity>=8.2.0

# Environment variable management
python-dotenv>=1.0.0

# Data processing (usually included with Python but explicit for clarity)
# Standard library modules used:
# - json (built-in)
# - os (built-in) 
# - sys (built-in)
# - datetime (built-in)
# - pathlib (built-in)
# - tempfile (built-in)
# - shutil (built-in)
# - typing (built-in)
# - enum (built-in)
# - re (built-in)
# - argparse (built-in)
# - asyncio (built-in)
# - time (built-in)
# - urllib.parse (built-in)
# - dataclasses (built-in, Python 3.7+)
